refactor: Debank
