+++ packages/web3-providers/src/CoinGecko/index.ts
@@ -1,3 +1,3 @@
-export * from './apis/CoinGecko.js'
+export * from './apis/TrendingAPI.js'
 export * from './apis/CoinGeckoEVM.js'
 export * from './apis/CoinGeckoSolana.js'
